Big refactor of handling of mesh data. De-interleaved vertex data (because it's a ton easier to work with).
Added helpers to enable common attributes like position, normal, indices and UVs. Better use of Buffer and VBO classes throughout.

Aborted attempt to move frustum to object space and only calculate the planes when the camera moves, which would save on calculations per frame, but it didn't work and it's late.