Victory, finally back to parity with pre-refactor.
Added a lights renderer. Phases of rendering are now a bit of a problem, though, since for example the light renderer needs to run before the meshes are rendered.

Can probably get away with explicit ordering for now, but might also be worthwhile to make it more explicit rather than having it configured in Main. Potentially just move it to WorldRenderer.