Fixed lights. Directional lights were using the world position instead of local.
Added debug options for lights back again, and added an icosphere rendering debug mode. Range might be nice to have as well.

Made it possible to add optional update data requests. You still have to have at least one requirement, though. Could change that to be "one of any of these components", but not sure there's a use case for it, so holding off for now.